TD-1555 Renames the Endpoint helper and directs to new tracker endpoint
