+++ DigitalLearningSolutions.Web/ViewModels/LearningMenu/DiagnosticContentViewModel.cs
@@ -36,7 +36,7 @@
 
             ContentSource = ContentViewerHelper.IsScormPath(diagnosticContent.DiagnosticAssessmentPath)
                 ? ContentViewerHelper.GetScormAssessmentSource(
-                    OldSystemEndpointHelper.GetScormPlayerUrl(config),
+                    SystemEndpointHelper.GetScormPlayerUrl(config),
                     centreId,
                     customisationId,
                     candidateId,
@@ -53,7 +53,7 @@
                     diagnosticContent.Version,
                     progressId,
                     type,
-                    OldSystemEndpointHelper.GetTrackingUrl(config),
+                    SystemEndpointHelper.GetTrackingUrl(config),
                     tutorials,
                     diagnosticContent.PassThreshold);
         }
